app/utils/config_loader.py
import yaml
from pathlib import Path


class ConfigLoader:
    @staticmethod
    def load_yaml(path: Path) -> dict:
        with open(path, "r", encoding="utf-8") as f:
            return yaml.safe_load(f)