requirements.txt
pyyaml
langchain
langchain-core
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pymupdf4llm
python-docx
dashscope
python-dotenv
fastapi
uvicorn
python-multipart
langgraph
sqlmodel